refactor: storage code
