--- src/common/modules/storage/deletePasswordFromStorage.ts
@@ -1,5 +1,0 @@
-import { deleteFromStorage } from '@common/modules/storage/deleteFromStorage'
-
-export async function deletePasswordFromStorage(groupId: number): Promise<void> {
-    await deleteFromStorage(`group-password-${groupId}`)
-}
