--- src/common/modules/storage/saveGroupIdsToStorage.ts
@@ -1,5 +1,0 @@
-import { saveToStorage } from '@common/modules/storage/saveToStorage'
-
-export async function saveGroupIdsToStorage(ids: number[]): Promise<void> {
-    await saveToStorage<number[]>('groupIds', ids)
-}
