--- src/common/modules/storage/savePasswordToStorage.ts
@@ -1,9 +1,0 @@
-import { saveToStorage } from '@common/modules/storage/saveToStorage'
-
-export async function savePasswordToStorage(
-    groupId: number,
-    pass: string,
-): Promise<void> {
-    const key = `group-password-${groupId}`
-    await saveToStorage<string>(key, pass)
-}
