+++ src/popup/components/Views/GroupView/EnterPassword.vue
@@ -10,7 +10,7 @@ import { useCryptoStore } from '@/stores/crypto'
 import { useRoute, useRouter } from 'vue-router'
 import { error } from '@common/modules/error'
 import { showToast } from '@common/modules/showToast'
-import { savePasswordToStorage } from '@common/modules/storage/savePasswordToStorage'
+import { savePasswordToStorage } from '@common/modules/storage/password'
 import ShieldCheckIcon from '@common/components/Icons/ShieldCheckIcon.vue'
 import Input from '@common/components/Form/Input.vue'
 import WarningBox from '@common/components/WarningBox.vue'
