+++ src/popup/components/Views/GroupView/IsUnlockedBox.vue
@@ -6,8 +6,7 @@ import { useGroupStore } from '@/stores/group'
 import { usePopupStore } from '@/stores/popup'
 import { useCryptoStore } from '@/stores/crypto'
 import { showToast } from '@common/modules/showToast'
-import { getPasswordFromStorage } from '@common/modules/storage/getPasswordFromStorage'
-import { deletePasswordFromStorage } from '@common/modules/storage/deletePasswordFromStorage'
+import { getPasswordFromStorage, deletePasswordFromStorage } from '@common/modules/storage/password'
 import LockClosedIcon from '@common/components/Icons/LockClosedIcon.vue'
 import WarningBox from '@common/components/WarningBox.vue'
 import SmallSpinner from '@common/components/SmallSpinner.vue'
