+++ src/popup/stores/attempts.ts
@@ -2,8 +2,7 @@ import type { Attempts } from '@/types'
 import { onMounted, ref } from 'vue'
 import { defineStore } from 'pinia'
 import { trans } from '@common/modules/trans'
-import { getFromStorage } from '@common/modules/storage/getFromStorage'
-import { saveToStorage } from '@common/modules/storage/saveToStorage'
+import { saveToStorage, getFromStorage } from '@common/modules/storage'
 import { error } from '@common/modules/error'
 import { showToast } from '@common/modules/showToast'
