+++ src/popup/stores/tabs.ts
@@ -7,7 +7,7 @@ import { showToast } from '@common/modules/showToast'
 import { getCurrentLinks } from '@/modules/tabs/getCurrentLinks'
 import { restoreTabs } from '@/modules/tabs/restoreTabs'
 import { closeTabs } from '@/modules/tabs/closeTabs'
-import { getPasswordFromStorage } from '@common/modules/storage/getPasswordFromStorage'
+import { getPasswordFromStorage } from '@common/modules/storage/password'
 
 export const useTabsStore = defineStore('tabs', () => {
     const groupStore = useGroupStore()
